requirements.txt
Flask==1.0.2
newsapi-python==0.2.6